renames diffractionimage.py -> virtualimage.py, misc. doc updates
